add discord presence plugin
This allows users to show off their currently playing song in
the popular messenger Discord. This requires the proprietary SDK
by Discord, which cannot be included here for obvious reasons.
But the SDK is available online for free, so packagers *can* get
it if they so desire.

Maybe in the future I'll find a way to provide better instructions
(or maybe even automation?) of this.

But for now: put the SDK wrapper (*.cpp and *.h) into discord/,
and put the SDK binaries (*.so etc.) into discord-lib/.
CMake handles the rest.

NOTE: I do not have mac OS / Windows available, so this is *untested*
on those systems, and likely to fail.